# app/public/openapi.yaml
openapi: 3.0.3
info:
  title: Nudlers Finance API
  description: |
    API for managing personal finance transactions, budgets, and bank/credit card scraping.
    
    ## Transaction Types
    Transactions can be of two types:
    - `credit_card` - Credit card transactions from vendors like visaCal, max, isracard, amex
    - `bank` - Bank account transactions from vendors like hapoalim, leumi, mizrahi, discount
    
    Use the `transactionType` query parameter to filter transactions by type.
  version: 1.0.0
  contact:
    name: Nudlers
  license:
    name: MIT
servers:
  - url: /api
    description: API Server
...
paths:
...
  /credentials/{id}:
    get:
      tags:
        - Integrations
      summary: Get a credential
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Credential details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Credential'
    put:
      tags:
        - Integrations
      summary: Update a credential
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialInput'
      responses:
        '200':
          description: Credential updated
    delete:
      tags:
        - Integrations
      summary: Delete a credential
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Credential deleted
...
components:
  schemas:
...
    Credential:
      type: object
      properties:
        id:
          type: integer
        vendor:
          type: string
          description: Bank or credit card vendor
        nickname:
          type: string
          description: Friendly name for this credential
        is_active:
          type: boolean
        last_synced_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    CredentialInput:
      type: object
      required:
        - vendor
      properties:
        vendor:
          type: string
          description: Bank or credit card vendor ID
        username:
          type: string
        password:
          type: string
        id_number:
          type: string
          description: ID number (for Israeli banks)
        card6_digits:
          type: string
          description: First 6 digits of card (for Isracard/Amex)
        nickname:
          type: string
        is_active:
          type: boolean
          default: true
...